Create longestSnake.cpp
You are given a 2-D array with same number of rows and columns. You have to determine the longest snake in the array. The property to find the snake is the difference between the adjacent(left, right, up or down) should be either 1 or -1. If there are more than one snakes with maximum length, the output should print both of them. 

Example--> 

The given array elements are as follows: 

4	7	9	8 
5	6	5	4 
6	7	8	5 
10	9	7	6 

The longest snakes are 7->6->5->4->5->6 and 7->6->7->6->5->4